server/api/requirements.txt
fastapi==0.109.0
python-multipart==0.0.9
python-dotenv==1.0.0
loguru==0.7.2
PyYAML==6.0.1
PyMuPDF==1.23.6
requests==2.31.0
Jinja2==3.1.2
tiktoken==0.5.1
uvicorn==0.27.0
slowapi==0.1.8
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.61.0
youtube-search==2.1.2
youtube-transcript-api==0.6.1
httpx==0.25.2
slugify==0.0.1
Pillow==10.0.1
cryptography==41.0.4
argon2-cffi==23.1.0